% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepl.R
\name{add_deepl_translation}
\alias{add_deepl_translation}
\title{Query DeepL translation for text items in list of translations that are NA}
\usage{
add_deepl_translation(
  trans.list = list(),
  target_languages = NULL,
  source_lang = "English",
  API_key = NULL
)
}
\arguments{
\item{trans.list}{List of translations as returned by \code{\link[=get_translations]{get_translations()}} or \code{\link[=update_translations]{update_translations()}}}

\item{target_languages}{Languages to be queried. By default uses names of \code{trans.list}}

\item{source_lang}{Source language, that is in which language questionnaire was designed. Default 'English'}

\item{API_key}{Authentication Key for DeepL API}
}
\value{
List of translations
}
\description{
Query DeepL translation for text items in list of translations that are NA
}
