% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepl.R
\name{add_deepl_translation_dt}
\alias{add_deepl_translation_dt}
\title{Add DeepL translation to data table. I.e. for an element of list of}
\usage{
add_deepl_translation_dt(
  dt,
  source_lang = "English",
  target_lang = NULL,
  API_key = NULL
)
}
\arguments{
\item{dt}{Data table of translations}

\item{source_lang}{Source language, that is in which language questionnaire was designed. Default 'English'}

\item{API_key}{Authentication Key for DeepL API}
}
\description{
Add DeepL translation to data table. I.e. for an element of list of
}
