% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_api.R
\name{add_gl_translate_dt}
\alias{add_gl_translate_dt}
\title{Add Google API  translation to data table. I.e. for an element of list of translations in tms}
\usage{
add_gl_translate_dt(
  dt,
  source_lang = "English",
  target_lang = NULL,
  API_key = NULL
)
}
\arguments{
\item{dt}{Data table of translations}

\item{source_lang}{Source language, that is in which language questionnaire was designed. Default 'English'}
}
\description{
Add Google API  translation to data table. I.e. for an element of list of translations in tms
}
