% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suso.R
\name{add_suso_sheet}
\alias{add_suso_sheet}
\title{Add provided user's Translation to a Survey Solutions Translation sheet}
\usage{
add_suso_sheet(
  questionnaire.list = "",
  sheet = "",
  translation.dt = "",
  pattern = NULL
)
}
\arguments{
\item{questionnaire.list}{List of Questionnaire Translation returned by \code{\link{get_suso_tfiles}}}

\item{sheet}{Sheet in \code{questionnaire.list} to which translation should be added}

\item{pattern}{Regular expression that matches question coding. Should be specified if \code{remove_coding()} was used to process translation file}

\item{data.table}{of translation to be added to Questionnaire}
}
\value{
data table of \code{sheet} with Translation added
}
\description{
Add provided user's Translation to a Survey Solutions Translation sheet
}
