% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_odk.R
\name{add_translation_odk}
\alias{add_translation_odk}
\title{Add translation to ODK Questionnaire File}
\usage{
add_translation_odk(
  tdb = list(),
  ss = NULL,
  sheets = c("survey", "choices"),
  translation.to.add = NULL,
  types = NULL
)
}
\arguments{
\item{tdb}{TODO}

\item{ss}{Identifier of a Google Sheet (see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}) that contains ODK questionnaire file}

\item{sheets}{TODO}

\item{translation.to.add}{TODO}

\item{types}{TODO}
}
\description{
Add translation to ODK Questionnaire File
}
