% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_api.R
\name{batchTranslate_GApi}
\alias{batchTranslate_GApi}
\title{Batch Translate Text Items Using Google Translate API}
\usage{
batchTranslate_GApi(
  tdb = list(),
  target_languages = NULL,
  source_lang = "English",
  auth = Sys.getenv("GL_AUTH")
)
}
\arguments{
\item{tdb}{List of translations as returned by \code{\link[=get_tdb_data]{get_tdb_data()}} or \code{\link[=update_tdbs]{update_tdbs()}}. This list should contain data tables with text items ready for translation.}

\item{target_languages}{Languages to be queried for translation. By default, if not specified, the function uses the names of the data tables in \code{tdb} as target languages. Specify target languages as a vector of language names or codes supported by Google Translate.}

\item{source_lang}{Source language, indicating the original language of the text items. The default is 'English'. Accurate specification is crucial for effective translations.}

\item{auth}{Authentication details for Google Translate API. This can be specified as the location of the service account credential JSON file. The default is obtained from the environment variable 'GL_AUTH'. For more details on setting up authentication, see \href{https://cloud.google.com/translate/docs/setup}{Google Translate API Setup}.}
}
\value{
Returns a modified list of data tables (\code{tdb}) where the text items have been translated to the specified target languages. Each data table in the list will have updated translations for the previously missing items.
}
\description{
This function \code{batchTranslate_GApi} interfaces with the Google Translate API to process a list of data tables (\code{tdb}), each containing text items for translation. It queries the Google Translate API for translations of text items that are missing (NA) in the 'Translation' column. The function is adaptable to multiple target languages and works on each data table in the list based on specified languages.
}
\examples{
\dontrun{
# Example usage (assuming 'tdb' is a predefined list of data tables):
translated_tdb <- batchTranslate_GApi(tdb, target_languages = c("German", "French"), auth = my_auth_file)
}

}
\seealso{
\code{gl_translate} in the \code{googleLanguageR} package for details on the underlying translation function.
\code{get_tdb_data()} and \code{update_tdbs()} for functions generating the input list of data tables.
}
