% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup.text.item}
\alias{cleanup.text.item}
\title{CLEANUP TEXT ITEM - REMOVE ONLY PROPER WHITESPACE TO NOT REMOVE TABS AND NEWLINES}
\usage{
cleanup.text.item(dt, regex = " ")
}
\description{
CLEANUP TEXT ITEM - REMOVE ONLY PROPER WHITESPACE TO NOT REMOVE TABS AND NEWLINES
}
