% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_titems}
\alias{collapse_titems}
\title{COLLAPSE dt OF TRANSLATION ITEMS INTO UNIQUE SET OF TRANSLATION ITEMS}
\usage{
collapse_titems(dt)
}
\arguments{
\item{dt}{}
}
\value{
data table of unique items
}
\description{
COLLAPSE dt OF TRANSLATION ITEMS INTO UNIQUE SET OF TRANSLATION ITEMS
}
