% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create.unique.var}
\alias{create.unique.var}
\title{CLEANUP TEXT ITEM - REMOVE ONLY PROPER WHITESPACE TO NOT REMOVE TABS AND NEWLINES}
\usage{
create.unique.var(dt, regex = " ", col = "value")
}
\description{
CLEANUP TEXT ITEM - REMOVE ONLY PROPER WHITESPACE TO NOT REMOVE TABS AND NEWLINES
}
