% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suso.R
\name{create_suso_file}
\alias{create_suso_file}
\title{Create Survey Solutions Translation file for 'Designer' Upload}
\usage{
create_suso_file(
  tdb.language,
  source_questionnaire = list(),
  path = stop("'path' must be specified"),
  sheets = NULL,
  statuses = c("machine", "reviewed", "translated"),
  qcode_pattern = NULL
)
}
\arguments{
\item{tdb.language}{data.table. Element (Language) of 'Translation Database' object returned by \code{\link{get_tdb_data}} or \code{\link{update_tdb}}}

\item{source_questionnaire}{data.table. 'Questionnaire Template' to which translation shall be added.  Usually an element of list returned by \code{\link{get_suso_tfiles}}}

\item{path}{Character. Writable file path where Translation File should be stored at, including file name and extension}

\item{sheets}{Character vector. For which sheets of questionnaire template file language will be added. Default all sheets that are found in template file}

\item{statuses}{Character vector. Which text items of which statuses from Translation Google Sheet should be merged? Matched without regard to case, surrounding spaces, or apostrophe variant.}

\item{qcode_pattern}{Regular expression that matches question coding. Should be specified if used in \code{parse_suso_titems()}.}
}
\description{
This function creates an .xlsx file based on the 'Translation Database' list and Survey Solutions source questionnaire.
The generated file can be uploaded to the Survey Solutions Designer.
}
\examples{
\dontrun{
# Take the 'German' Translation from database and merge into the source
questionnaire as returned by \code{\link{get_suso_tfiles}}
# Consider only text items of particular statuses defined by Translator
create_suso_file(
  tdb.language = new_tdb[["German"]],
  source_questionnaire = suso_trans_templates[["NAME-OF-QUESTIONNAIRE"]],
  statuses = c("Machine", "reviewed", "translated"),
  path = "your-path/German_NAME-OF-QUESTIONNAIRE.xlsx"
)
}

}
