% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suso.R
\name{create_suso_sheet}
\alias{create_suso_sheet}
\title{Add provided user's Translation to a Survey Solutions Translation sheet}
\usage{
create_suso_sheet(
  source_questionnaire = "",
  sheet = "",
  language.tdb.dt = "",
  pattern = NULL
)
}
\arguments{
\item{source_questionnaire}{List. 'Questionnaire Template' to which translation shall be added.  Usually an element of list returned by \code{\link{get_suso_tfiles}}}

\item{sheet}{Sheet in \code{source_questionnaire} to which translation should be added}

\item{pattern}{Regular expression that matches question coding. Should be specified if \code{remove_coding()} was used to process translation file}

\item{data.table}{of translation to be added to Questionnaire}
}
\value{
data table of \code{sheet} with Translation added
}
\description{
Add provided user's Translation to a Survey Solutions Translation sheet
}
