% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk.R
\name{get_odk_wsheet}
\alias{get_odk_wsheet}
\title{Pull data from ODK Google Sheet}
\usage{
get_odk_wsheet(gs = "", sheets = c("survey", "choices"))
}
\arguments{
\item{gs}{Identifier of a Google Sheet, see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}}

\item{sheets}{Character vector of sheets to be read from \code{gs}}
}
\value{
List of data.tables which are the respective sheets of worksheet
}
\description{
Pulls sheets from google sheet on which the questionnaire is developed.
}
