% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suso.R
\name{get_sursol_titems}
\alias{get_sursol_titems}
\title{Read SurSol Translation File}
\usage{
get_sursol_titems(
  questionnaires = NULL,
  user = "",
  password = "",
  sheets = NULL,
  types = c("Title", "Instruction", "OptionTitle", "ValidationMessage", "SpecialValue")
)
}
\arguments{
\item{questionnaires}{Named character vector of instruments to source. Elements must be id of questionnaire. Name is the Title of Questionnaire.}

\item{user}{SuSo User Name}

\item{password}{SuSo Password}

\item{sheets}{Sheets from Translation to read. Default all}

\item{types}{Which type of text items to keep}
}
\value{
Data table of unique text items in instrument(s)
}
\description{
Read SurSol Translation File
}
