% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suso.R
\name{get_sursol_titems_byqx}
\alias{get_sursol_titems_byqx}
\title{Get unique Text Items of questionnaire}
\usage{
get_sursol_titems_byqx(
  qxid = NULL,
  user = "",
  password = "",
  sheets = NULL,
  types = c("Title", "Instruction", "OptionTitle", "ValidationMessage", "SpecialValue")
)
}
\arguments{
\item{qxid}{Id of Questionnaire in SuSo Designer}

\item{user}{SuSo Designer User Name}

\item{password}{SuSo Designer Password}

\item{sheets}{Sheets from Translation to read. Default all}

\item{types}{Which type of text items to keep}
}
\description{
Get unique Text Items of questionnaire
}
