% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suso.R
\name{get_suso_tfile}
\alias{get_suso_tfile}
\title{Retrieve translation file from SurSol Designer}
\usage{
get_suso_tfile(questionnaire = "", user = "", password = "", sheets = NULL)
}
\arguments{
\item{user}{SuSo Designer User Name}

\item{password}{SuSo Designer Password}

\item{sheets}{Sheets from Translation to read. Default all}

\item{qx_id}{Questionnaire id within Survey Solutions Designer}
}
\value{
List of sheets in translation file
}
\description{
Retrieve translation file from SurSol Designer
}
