% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_suso.R
\name{get_suso_tfiles}
\alias{get_suso_tfiles}
\title{Retrieve mutiple translation files from SurSol Designer}
\usage{
get_suso_tfiles(questionnaires = "", user = "", password = "", sheets = NULL)
}
\arguments{
\item{questionnaires}{Vector of Questionnaire id's within Survey Solutions Designer}

\item{user}{SuSo Designer User Name}

\item{password}{SuSo Designer Password}

\item{sheets}{Sheets from Translation to read. Default all}
}
\value{
List of translation file with list of sheets
}
\description{
Wrapper for \code{\link{get_suso_tfile}} to retrieve multiple translation files
}
