% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_software_mismatch.R
\name{get_sw_issues_dt}
\alias{get_sw_issues_dt}
\title{Check individual sheet (data.table) for software mismatches}
\usage{
get_sw_issues_dt(dt, pattern = "\%[a-zA-Z0-9_]+\%", lang = NULL)
}
\arguments{
\item{dt}{data table of a translation}

\item{pattern}{pattern that identifies text item in Original/Translation (e.g. text substitution '\%rostertitle\%')}

\item{lang}{name of sheet/language. for message}
}
\value{
data table with text items that contain issues
}
\description{
Check individual sheet (data.table) for software mismatches
}
