% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdb.R
\name{get_tdb_data}
\alias{get_tdb_data}
\title{Pull 'Translation Database' Google Sheet}
\usage{
get_tdb_data(ss = "", sheets = NULL)
}
\arguments{
\item{ss}{Identifier of a Google Sheet, see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}}

\item{sheets}{Sheets in ss to read. Should be ISO language names. By default reads all sheets found in \code{ss}}
}
\value{
A list object where each element is one translation sheet
}
\description{
A simple wrapper for \code{\link[googlesheets4]{read_sheet}}.
Retrieves all data from the 'Translation Database' Google Sheet.
}
\examples{
\dontrun{
get_tdb_data(ss = "GOOGLE-SHEET-IDENTIFIER")
}
}
