% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{get_translations}
\alias{get_translations}
\title{Read Translation Sheets from Master Worksheet}
\usage{
get_translations(gs = "", sheets = NULL)
}
\arguments{
\item{gs}{Identifier of a Google Sheet, see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}}

\item{sheets}{Sheets in gs to read. Should be ISO languages. By default reads all sheets found in \code{gs}}
}
\value{
list of sheets
}
\description{
Read Translation Sheets from Master Worksheet
}
