% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_translation.R
\name{lookup_translation}
\alias{lookup_translation}
\title{Lookup Translations for Untranslated Text Items}
\usage{
lookup_translation(dt, min.dist = 0.2)
}
\arguments{
\item{dt}{A data.table object containing text items and translations for a specific language.}

\item{min.dist}{The minimum distance threshold. Text items with similarity distance greater than this threshold will not get a matched reference. The closer the number is to 0, the more similar the two strings. If the number is 1, the strings are completely different.}
}
\value{
A modified data.table with additional column 'Similarity' for matched reference translations and similarity details.
}
\description{
This function identifies text items without translations and looks up the closest match in the reference translations.
}
\examples{
  \dontrun{
# Assuming `new_tdb` is a list of data.tables for different languages as returned by `update_tdb()`
lookup_translation(new_tdb$Portuguese)
}
}
