% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk.R
\name{parse_odk_titems}
\alias{parse_odk_titems}
\title{Parse translation items returned from 'get_odk_titems' into clean data table}
\usage{
parse_odk_titems(dt)
}
\arguments{
\item{dt}{Data table returned by get_odk_titems}
}
\value{
Data table
}
\description{
Parse translation items returned from 'get_odk_titems' into clean data table
}
