% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suso.R
\name{parse_sursol_titems}
\alias{parse_sursol_titems}
\title{Get (unique) Text Items from list of Survey Solutions translation(s) in rectangular format}
\usage{
parse_sursol_titems(
  translation,
  sheets = NULL,
  types = c("Title", "Instruction", "OptionTitle", "ValidationMessage", "SpecialValue",
    "FixedRosterTitle"),
  collapse = TRUE
)
}
\arguments{
\item{translation}{Named list where each element contains the translation sheets of each instrument}

\item{sheets}{Sheets from Translation to read. Default all}

\item{types}{Which type of text items to keep}

\item{collapse}{Boolean. If TRUE, only 'unique' text items will be returned.}
}
\value{
Data table of unique text items in instrument(s)
}
\description{
Get (unique) Text Items from list of Survey Solutions translation(s) in rectangular format
}
