% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_suso.R
\name{parse_suso_titems}
\alias{parse_suso_titems}
\title{Parse 'Source Questionnaires' object into unique text items}
\usage{
parse_suso_titems(
  tmpl_list,
  sheets = NULL,
  types = c("Title", "Instruction", "OptionTitle", "ValidationMessage", "SpecialValue",
    "FixedRosterTitle", "CriticalRuleMessage"),
  collapse = TRUE,
  qcode_pattern = NULL
)
}
\arguments{
\item{tmpl_list}{A named nested list as returned by \code{\link{get_suso_tfiles}} (Source Questionnaire object)}

\item{sheets}{A character vector of sheet names within \code{tmpl_list} files to be parsed. By default, all sheets are parsed.}

\item{types}{A character vector specifying the types of text items to keep.
By default, includes \code{CriticalRuleMessage}. Supply a subset to exclude types.}

\item{collapse}{Boolean. If TRUE, only unique text items will be returned.}

\item{qcode_pattern}{Optional. Regex pattern to remove question codes, e.g., "Q1." in "Q1. How old are you?". Default is NULL, keeping texts unmodified.}
}
\value{
A data.table containing (unique) text items found across questionnaires in \code{tmpl_list}.
}
\description{
This function processes a nested listed returned by \code{\link{get_suso_tfiles}}.
It identifies all text items across all sheets across all questionnaires found in \code{tmpl_list}. If \code{collapse=TRUE} only unique text items will be returned.
}
\examples{
\dontrun{
source_titems <- parse_suso_titems(
tmpl_list = suso_trans_templates, #Nested list as returned by `get_suso_tfiles()`
collapse = TRUE # Keep only unique text items across questionnaires
)
}
}
