% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tms}
\alias{setup_tms}
\title{Creates new 'Translation Master' Google Sheet}
\usage{
setup_tms(
  ssheet_name = stop("'ssheet_name' must be specified"),
  lang_names = stop("'lang_names' must be specified")
)
}
\arguments{
\item{ssheet_name}{Character. The name that will be assigned to the Google Sheet.}

\item{lang_names}{Character Vector. A vector of \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639} language names for which translation will be needed. Each language becomes one sheet/tab.}
}
\value{
List of metadata on the newly created sheet.
}
\description{
Creates a new Google Sheet (based on the \href{https://docs.google.com/spreadsheets/d/1xLHEDm5bgtv4IHHCbfyMf_sdNsdzz_AUcRrnpb2fXTw}{translation template}) in your Google user account. This new sheet will serve as the main 'Translation Master' Google Sheet in subsequent workflow.
}
\examples{
\dontrun{
# Create a new translation master sheet with two languages (German and French)
new_sheet <- setup_tms(
  ssheet_name = "Project X: Translation Sheet",
  lang_names = c("German", "French")
)

# The returned 'new_sheet' object contains metadata about the created sheet
new_sheet$spreadsheet_url # Access the URL of the created sheet

# Print all metadata
new_sheet
}

}
