% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tsheet}
\alias{setup_tsheet}
\title{Creates new google worksheet using translation template}
\usage{
setup_tsheet(
  name.ssheet = stop("'name.ssheet' must be specified"),
  translation.languages = stop("'translation.languages' must be specified")
)
}
\arguments{
\item{name.ssheet}{Character. Name that will be assigned to  Google Worksheet}

\item{translation.languages}{Character Vector. ISO 639-1 languages for which translation will be needed. Each language becomes one sheet}
}
\value{
List of metadata on new sheet
}
\description{
Creates new google worksheet using translation template
}
