% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{update_translation}
\alias{update_translation}
\title{Compares current questionnaire file against list of existing translations}
\usage{
update_translation(curr.trans = list(), new.items = data.table())
}
\arguments{
\item{curr.trans}{List of translations as returned by \code{\link[=get_translations]{get_translations()}}}

\item{new.items}{Data table of questionnaire text items returned by either \code{\link[=parse_odk_titems]{parse_odk_titems()}} or \code{\link[=get_sursol_titems]{get_sursol_titems()}}}
}
\value{
List of updated translations
}
\description{
Removes any text item that no longer is part of the questionnaire(s).
Adds any text item that was not part of translation before
}
