% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{update_translation_lement}
\alias{update_translation_lement}
\title{Update translation for one element in list}
\usage{
update_translation_lement(curr.trans, new.items)
}
\arguments{
\item{curr.trans}{List of translations as returned by \code{\link[=get_translations]{get_translations()}}}

\item{new.items}{Data table of questionnaire text items returned by either \code{\link[=parse_odk_titems]{parse_odk_titems()}} or \code{\link[=get_sursol_titems]{get_sursol_titems()}}}
}
\description{
Update translation for one element in list
}
