% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdb.R
\name{write_tdb_data}
\alias{write_tdb_data}
\title{Write one data.table to 'Translation Database' Google Sheet}
\usage{
write_tdb_data(dt, ss = "", sheet = "", range = "A2:G", col_names = FALSE)
}
\arguments{
\item{dt}{A data.table to write. Usually an element within the list of translations
returned by \code{update_tdb()} that contains the translation of one language.}

\item{ss}{Identifier of a Google Sheet, see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}}

\item{sheet}{The name of the sheet in \code{ss}. The ISO 639 language name, which
is the name of the element in the list of translations.}

\item{range}{Cell Range to write}

\item{col_names}{A logical value indicating whether to send the column names of the data.}
}
\description{
A simple wrapper for \code{\link[googlesheets4:range_write]{googlesheets4::range_write()}}. Writes a data.table to the specified
Google Sheet and sheet name.
}
\examples{
\dontrun{
#Write one particular (updated) language to the 'Translation Database' Google Sheet
write_tdb_data(new_tdb[["German"]],
               ss = "GOOGLE-SHEET-IDENTIFIER",
               sheet = "German"
)

#Or write all languages found in 'Translation Database' object to the Google Sheet
purrr::walk(
  .x = names(new_tdb),
  .f = ~ write_tdb_data(new_tdb[[.x]],
                        ss = "GOOGLE-SHEET-IDENTIFIER",
                        sheet = .x
  )
)
}
}
