% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suso.R
\name{add_suso_sheet}
\alias{add_suso_sheet}
\title{Add provided user's Translation to a Survey Solutions Translation sheet}
\usage{
add_suso_sheet(file = "", sheet = "", translation.file = "", pattern = NULL)
}
\arguments{
\item{file}{Excel file (SuSo Translation File)}

\item{sheet}{Sheet in \code{file} to which translation should be added}
}
\value{
data table of \code{sheet} with Translation added
}
\description{
Add provided user's Translation to a Survey Solutions Translation sheet
}
