% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odk_wsheet.R
\name{get_odk_wsheet}
\alias{get_odk_wsheet}
\title{Pull data from ODK Google Sheet}
\usage{
get_odk_wsheet(wsid = "", sheets = c("survey", "choices"))
}
\arguments{
\item{wsid}{Identifier of a Google Work Sheet}

\item{sheets}{Character vector of sheets to be read on `wisid`}
}
\value{
List of data.tables which are the respective sheets of worksheet
}
\description{
Pulls sheets from google sheet on which the questionnaire is developed.
}
