% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_coding.R
\name{remove_coding}
\alias{remove_coding}
\title{Remove the (question) coding of a text item (usually 'Question') from set of translations}
\usage{
remove_coding(dt, pattern = "", collapse = TRUE)
}
\arguments{
\item{dt}{Data table of questionnaire text items returned by either \code{\link[=parse_odk_titems]{parse_odk_titems()}} or \code{\link[=get_sursol_titems]{get_sursol_titems()}}}

\item{pattern}{Regular expression that matches question coding}

\item{collapse}{boolean to indicate if after removing question coding, the set of translation items should be scanned for/collapsed to unique items}
}
\value{
dt
}
\description{
Remove the (question) coding of a text item (usually 'Question') from set of translations
}
