% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_check.R
\name{syntax_check}
\alias{syntax_check}
\title{Identifies software-related mismatches between 'Text_Item' and 'Translation'}
\usage{
syntax_check(tdb = list(), pattern = "\%[a-zA-Z0-9_]+\%")
}
\arguments{
\item{tdb}{List. Translations created by \code{get_tdb_data()} or \code{update_tdb()}}

\item{pattern}{Character. Regular expression that identifies text item in Original/Translation (e.g. text substitution '\%rostertitle\%')}

\item{languages}{Character Vector. For which sheets/languages shall software-related mismatches be checked?}
}
\value{
List of translations as supplied by user, with software mismatches flagged
}
\description{
Updates column 'Status' and leaves a 'Comment/Note' that explains which issue is identified.
}
