% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{write_trans_tsheet}
\alias{write_trans_tsheet}
\title{Write a data table to Translation Sheet
Simple wrapper for \code{\link[googlesheets4:range_write]{googlesheets4::range_write()}}}
\usage{
write_trans_tsheet(dt, gs = "", sheet = "", range = "A2:G", col_names = FALSE)
}
\arguments{
\item{dt}{Data table to be written}

\item{gs}{Identifier of a Google Sheet, see \code{\link[googlesheets4:sheets_id]{googlesheets4::sheets_id()}}}

\item{sheet}{Name of sheet in \code{gs}}

\item{range}{Cell Range to write}

\item{col_names}{Logical, indicates whether to send the column names of data}
}
\description{
Write a data table to Translation Sheet
Simple wrapper for \code{\link[googlesheets4:range_write]{googlesheets4::range_write()}}
}
